pyproject.toml
[tool.poetry]
name = "llm-recepies"
version = "0.1.0"
description = "Project helping with playing with different LLM's"
authors = ["ksopyla <krzysztofsopyla@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "llm_recepies"}]

[tool.poetry.dependencies]
python = "~3.10"
torch = "^2.0.1"
transformers = "^4.30.1"
langchain = "^0.0"
openai = "^0.27.8"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.23.1"
ruff = "^0.0.272"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
